No need to NULL-terminate array of known size.
For some reason (which I yet do not understand) re2c uses size of DFA
state's "kernel" to determine whether a new state should be constructed
or some already constructed state (of the same "kernel" size) will do.

Since re2c knows "kernel" size for each state anyway, there's no need
to NULL-terminate "kernel" array while iterating over it: we can use
pre-calculated size instead.

+++ re2c/src/ir/dfa/state.h
@@ -17,7 +17,7 @@ public:
 	State * next;
 	State * link;
 	uint32_t depth; // for finding SCCs
-	ptrdiff_t kCount;
+	uint32_t kCount;
 	Ins ** kernel;
 
 	bool isPreCtxt;
